Detach Vulnerabilities into VulnerabilitiesMixin, filter->filter_string
