+++ gvm/protocols/gmpv208/__init__.py
@@ -135,6 +135,7 @@ from gvm.protocols.gmpv208.entities.users import (
     UsersMixin,
     get_user_auth_type_from_string,
 )
+from gvm.protocols.gmpv208.entities.vulnerabilities import VulnerabilitiesMixin
 from gvm.connections import GvmConnection
 
 PROTOCOL_VERSION = (20, 8)
@@ -171,6 +172,7 @@ class Gmp(
     SecInfoMixin,
     UserSettingsMixin,
     UsersMixin,
+    VulnerabilitiesMixin,
 ):
     def __init__(
         self,
